Transfer all work back to db_seed.js
