+++ DB/seeders/db_seed.js
@@ -1,63 +1,38 @@
-require('dotenv').config(); // Allows pool to recognize enviroment variables. Gotta be before pool variable definition!!!
+require('dotenv').config();
 const faker = require('faker');
 const moment = require('moment');
+const { Seeder } = require('mysql-db-seed');
 const pool = require('../../connection/connection');
 
-pool.getConnection((err, connection) => {
-  try {
-    [...Array(100)].forEach(() => {
-      connection.query(
-        `INSERT INTO countries (name) VALUES ("${faker.address.country()}")`,
-        () => {},
-      );
-    });
-    console.log('Inserted into countries');
-  } catch (error) {
-    if (error) console.error(`Error: ${error}`);
+const seed = new Seeder(
+  10,
+  process.env.DB_HOST,
+  process.env.DB_USER,
+  process.env.DB_PASS,
+  process.env.DB_NAME,
+);
+
+// Function to determin whether a country is called
+// British Indian Ocean Territory (Chagos Archipelago) or not
+// becuase the name is too long to fit in varchar(45)
+function isNotBIO() {
+  let country = faker.address.country();
+
+  while (country === 'British Indian Ocean Territory (Chagos Archipelago)') {
+    country = faker.address.country();
   }
-});
+
+  return country;
+}
 
 pool.getConnection((err, connection) => {
   try {
-    [...Array(50)].forEach(() => {
-      connection.query(
-        `INSERT INTO states (name) VALUES ("${faker.address.state()}")`,
-        () => {},
-      );
-    });
-    console.log('Inserted into states');
-  } catch (error) {
-    if (error) console.error(`Error: ${error}`);
-  }
-});
-
-pool.getConnection((err, connection) => {
-  try {
-    [...Array(100)].forEach(() => {
-      connection.query(
-        `INSERT INTO addresses(
-            street,
-            street_no,
-            region,
-            zipcode,
-            country_id,
-            state_id
-            ) VALUES (
-            "${faker.address.streetName()}",
-            "${faker.random.number({ min: 10000, max: 99999 })}",
-            "region",
-            "${faker.address.zipCode()}",
-            "${faker.random.number({ min: 1, max: 100 })}",
-            "${faker.random.number({ min: 1, max: 50 })}"
-            )
-            `,
-        (error, results) => {
-          if (error) console.error(`Error: ${error}`);
-          console.log(results);
-        },
-      );
-    });
-    console.log('Inserted into addresses');
+    connection.query(
+      'INSERT INTO roles (name) VALUES ("admin"), ("user");',
+      () => {
+        connection.release();
+      },
+    );
   } catch (error) {
     if (error) console.error(`Error: ${error}`);
   }
@@ -66,158 +41,77 @@ pool.getConnection((err, connection) => {
 pool.getConnection((err, connection) => {
   try {
     connection.query(
-      'INSERT INTO genders (name) VALUES ("male"), ("female"), ("other")',
-      () => {},
+      'INSERT INTO genders (name) VALUES ("male"), ("female"), ("other");',
+      () => {
+        connection.release();
+      },
     );
-    console.log('Inserted into genders');
   } catch (error) {
     if (error) console.error(`Error: ${error}`);
   }
 });
 
-pool.getConnection((err, connection) => {
-  try {
-    [...Array(100)].forEach(() => {
-      connection.query(
-        `INSERT INTO patients(
-          first_name,
-          last_name,
-          birth_date,
-          telephone,
-          mobile,
-          amka,
-          afm,
-          comments,
-          gender_id,
-          address_id
-        ) VALUES (
-          "${faker.name.firstName()}",
-          "${faker.name.lastName()}",
-          "${moment(faker.date.past()).format('YYYY-MM-DD')}",
-          "${faker.phone.phoneNumber()}",
-          "${faker.phone.phoneNumber()}",
-          "${faker.random.number({ min: 10000, max: 99999 })}",
-          "${faker.random.number({ min: 10000, max: 99999 })}",
-          "${faker.lorem.sentence()}",
-          "${faker.random.number({ min: 1, max: 3 })}",
-          "${faker.random.number({ min: 1, max: 100 })}"
-        )`,
-        (error, results) => {
-          if (error) console.error(`Error: ${error}`);
-          console.log(results);
-        },
-      );
-    });
-    console.log('Inserted into patients');
-  } catch (error) {
-    if (error) console.error(`Error: ${error}`);
-  }
-});
+(async () => {
+  await seed.seed(99, 'countries', {
+    name: isNotBIO,
+  });
 
-pool.getConnection((err, connection) => {
-  try {
-    [...Array(100)].forEach(() => {
-      connection.query(
-        `INSERT INTO professions (name) VALUES ("${faker.name.jobTitle()}")`,
-        () => {},
-      );
-    });
-    console.log('Inserted into professions');
-  } catch (error) {
-    if (error) console.error(`Error: ${error}`);
-  }
-});
+  await seed.seed(49, 'states', {
+    name: faker.address.state,
+  });
 
-pool.getConnection((err, connection) => {
-  try {
-    connection.query(
-      'INSERT INTO roles (name) VALUES ("admin"), ("user")',
-      () => {},
-    );
-    console.log('Inserted into roles');
-  } catch (error) {
-    if (error) console.error(`Error: ${error}`);
-  }
-});
+  await seed.seed(99, 'addresses', {
+    street: faker.address.streetName,
+    street_no: () => faker.random.number({ min: 1, max: 30 }),
+    region: 'region',
+    zipcode: faker.address.zipCode,
+    country_id: () => faker.random.number({ min: 1, max: 100 }),
+    state_id: () => faker.random.number({ min: 1, max: 50 }),
+  });
 
-pool.getConnection((err, connection) => {
-  try {
-    [...Array(100)].forEach(() => {
-      connection.query(
-        `
-        INSERT INTO users(
-          first_name,
-          last_name,
-          email,
-          password,
-          birth_date,
-          created_at,
-          amka,
-          afm,
-          role_id,
-          profession_id
-        ) VALUES (
-          "${faker.name.firstName()}",
-          "${faker.name.lastName()}",
-          "${faker.internet.email()}",
-          "${faker.internet.password()}",
-          "${moment(faker.date.past()).format('YYYY-MM-DD')}",
-          "${moment(faker.date.past()).format('YYYY-MM-DD')}",
-          "${faker.random.number({ min: 10000, max: 99999 })}",
-          "${faker.random.number({ min: 10000, max: 99999 })}",
-          "${faker.random.number({ min: 1, max: 2 })}",
-          "${faker.random.number({ min: 1, max: 100 })}"
-        )`,
-        () => {},
-      );
-    });
-    console.log('Inserted into users');
-  } catch (error) {
-    if (error) console.error(`Error: ${error}`);
-  }
-});
+  await seed.seed(99, 'professions', {
+    name: faker.name.jobTitle,
+  });
 
-pool.getConnection((err, connection) => {
-  try {
-    [...Array(100)].forEach(() => {
-      connection.query(
-        `
-        INSERT INTO visits(
-          date,
-          comment,
-          start_time,
-          end_time,
-          user_id
-        ) VALUES (
-          "${moment(faker.date.past()).format('YYYY-MM-DD')}",
-          "${faker.lorem.sentence()}",
-          "${moment(faker.date.recent()).format('HH:MM:SS')}",
-          "${moment(faker.date.recent()).format('HH:MM:SS')}",
-          "${faker.random.number({ min: 1, max: 100 })}"
-        )`,
-        () => {},
-      );
-    });
-    console.log('Inserted into visits');
-  } catch (error) {
-    if (error) console.error(`Error: ${error}`);
-  }
-});
+  await seed.seed(99, 'patients', {
+    first_name: faker.name.firstName,
+    last_name: faker.name.lastName,
+    birth_date: () => moment(faker.date.past()).format('YYYY-MM-DD'),
+    telephone: faker.phone.phoneNumber,
+    mobile: faker.phone.phoneNumber,
+    amka: () => faker.random.number({ min: 1000000000, max: 9999999999 }),
+    afm: () => faker.random.number({ min: 1000000000, max: 9999999999 }),
+    comments: faker.lorem.sentence,
+    gender_id: () => faker.random.number({ min: 1, max: 3 }),
+    address_id: () => faker.random.number({ min: 1, max: 100 }),
+  });
 
-pool.getConnection((err, connection) => {
-  try {
-    [...Array(100)].forEach(() => {
-      connection.query(
-        `INSERT INTO users_patients(
-          user_id,
-          patient_id
-          ) VALUES (
-            "${faker.random.number({ min: 1, max: 100 })}",
-            "${faker.random.number({ min: 1, max: 100 })}"
-          )`,
-      );
-    });
-  } catch (error) {
-    if (error) console.error(`Error: ${error}`);
-  }
-});
+  await seed.seed(99, 'users', {
+    first_name: faker.name.firstName,
+    last_name: faker.name.lastName,
+    email: faker.internet.email,
+    password: faker.internet.password,
+    birth_date: () => moment(faker.date.past()).format('YYYY-MM-DD'),
+    created_at: () => seed.nativeTimestamp(),
+    amka: () => faker.random.number({ min: 1000000000, max: 9999999999 }),
+    afm: () => faker.random.number({ min: 1000000000, max: 9999999999 }),
+    role_id: () => faker.random.number({ min: 1, max: 2 }),
+    profession_id: () => faker.random.number({ min: 1, max: 100 }),
+  });
+
+  await seed.seed(99, 'visits', {
+    date: () => moment(faker.date.recent()).format('YYYY-MM-DD'),
+    comment: faker.lorem.sentence,
+    start_time: () => `${moment(faker.date.recent()).format('HH:MM')}:00`,
+    end_time: () => `${moment(faker.date.recent()).format('HH:MM')}:00`,
+    user_id: () => faker.random.number({ min: 1, max: 100 }),
+  });
+
+  await seed.seed(99, 'users_patients', {
+    user_id: () => faker.random.number({ min: 1, max: 100 }),
+    patient_id: () => faker.random.number({ min: 1, max: 100 }),
+  });
+
+  seed.exit();
+  process.exit();
+})();
